requirements.txt
fastapi
uvicorn
pydantic
typing
requests
black
gradio
pylint
joblib
pyyaml
pytest
sklearn
hydra-core
